pyproject.toml
[build-system]
requires = ["maturin>=1,<2"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]

[tool.uv]
cache-keys = [
  { file = "pyproject.toml" },
  { file = "pyro_mysql/*.pyi" }, # typing stub
  { file = "pyro_mysql/*.py" },  # typing stub
  { file = "src/**/*.rs" },
]

[project]
name = "pyro-mysql"
description = "High-performance MySQL driver for Python"
readme = "README.md"
license = { text = "MIT" }
keywords = ["mysql", "async", "asyncio", "database", "rust", "pyo3"]
dynamic = ["version"]
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Rust",
  "Topic :: Database",
  "Topic :: Database :: Front-Ends",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Framework :: AsyncIO",
  "Typing :: Typed",
]
dependencies = ["ty>=0.0.1a21"]

[project.urls]
Repository = "https://github.com/elbaro/pyro-mysql"

[dependency-groups]
dev = [
  "maturin>=1.9.4",
  "pdoc>=15.0.4",
  "pyright>=1.1.405",
  "pytest-asyncio>=1.2.0",
  "pytest-benchmark>=5.1.0",
  "rich>=14.1.0",
  "aiomysql>=0.2.0",
  "asyncmy>=0.2.10",
  "mysqlclient>=2.2.7",
  "pymysql>=1.1.2",
]
